Refactor and consolidate engine buffer architecture
- Centralized all buffer-related classes (SnapshotBuffer, ArenaAllocator, BitWriter/Reader, etc.) into Engine/Shared/Buffers/.
- Introduced IBuffer and ISlabAllocator for standard memory management and diagnostic tracking.
- Migrated all buffer positions and capacities to 'long' to support large data snapshots.
- Optimized bit-level serialization with byte-aligned fast paths and support for ReadOnlySequence<byte>.
- Fixed architectural flaws in virtual indexing and cross-segment patching.
- Added comprehensive unit tests in BufferArchitectureTests.cs.

+++ Engine/Core/Services/GameStateSnapshotter.cs
@@ -103,7 +103,7 @@ namespace Core
                             {
                                 var writer = new BitWriter(buffer);
                                 _binarySnapshotService.SerializeBitPackedDelta(ref writer, batch, null);
-                                segmentTasks[b] = buffer.AsSpan(0, writer.BytesWritten).ToArray();
+                                segmentTasks[b] = buffer.AsSpan(0, (int)writer.BytesWritten).ToArray();
                             }
                             finally { ArrayPool<byte>.Shared.Return(buffer); }
                         }).GetAwaiter().GetResult();
@@ -130,7 +130,7 @@ namespace Core
                             var array = new GameObject[actualCount];
                             Array.Copy(objects, 0, array, 0, actualCount);
                             _binarySnapshotService.SerializeBitPackedDelta(ref writer, array, null);
-                            return buffer.AsSpan(0, writer.BytesWritten).ToArray();
+                            return buffer.AsSpan(0, (int)writer.BytesWritten).ToArray();
                         }
                         catch (IndexOutOfRangeException)
                         {
@@ -187,7 +187,7 @@ namespace Core
                     {
                         var writer = new BitWriter(buffer);
                         _binarySnapshotService.SerializeBitPackedDelta(ref writer, objects, null);
-                        return buffer.AsSpan(0, writer.BytesWritten).ToArray();
+                        return buffer.AsSpan(0, (int)writer.BytesWritten).ToArray();
                     }
                     catch (IndexOutOfRangeException)
                     {
@@ -215,7 +215,7 @@ namespace Core
                     {
                         var writer = new BitWriter(buffer);
                         _binarySnapshotService.SerializeBitPackedDelta(ref writer, objects, null);
-                        return buffer.AsSpan(0, writer.BytesWritten).ToArray();
+                        return buffer.AsSpan(0, (int)writer.BytesWritten).ToArray();
                     }
                     catch (IndexOutOfRangeException)
                     {
